requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.94.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
JayDeBeApi==1.2.3
Jinja2==3.1.2
JPype1==1.4.1
MarkupSafe==2.1.2
packaging==23.0
pydantic==1.10.6
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.0.post1
typing_extensions==4.5.0
uvicorn==0.21.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.3
zipp==3.15.0
